+++ src/components/common/index.tsx
@@ -3,4 +3,3 @@ export { default as Header } from "./Header";
 export { default as NavigationBar } from "./NavigationBar";
 export { default as PostCard } from "./PostCard/PostCard";
 export { default as Background } from "./Background";
-export { default as ContextMenu } from "./ContextMenu";
